[feat] add multiple schedule update

+++ src/app/org-schedule/org-schedule.service.ts
@@ -2,9 +2,10 @@ import { Injectable } from '@nestjs/common';
 import { UpdateOrgScheduleDto } from './dto/update-org-schedule.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { OrgSchedule } from './entities/org-schedule.entity';
-import { Repository } from 'typeorm';
+import { In, Repository } from 'typeorm';
 import { OnEvent } from '@nestjs/event-emitter';
 import { defaultScheduleData } from '@/utils/data/org-schedule.data';
+import { UpdateMultiScheduleDto } from './dto/update-many.dto';
 
 @Injectable()
 export class OrgScheduleService {
@@ -36,13 +37,34 @@ export class OrgScheduleService {
   update(id: number, updateOrgScheduleDto: UpdateOrgScheduleDto) {
     const { startTime, endTime } = updateOrgScheduleDto;
     const newSchedule = this.orgScheduleRepository.create({
+      id,
       startTime,
       endTime,
     });
     return this.orgScheduleRepository.save(newSchedule);
   }
 
-  remove(id: number) {
+  async updateMany(
+    orgId: number,
+    updateMultiScheduleDto: UpdateMultiScheduleDto,
+  ) {
+    const { schedules } = updateMultiScheduleDto;
+    const ids = schedules.map((schedule) => schedule.id);
+    await this.checkOwnership(ids, orgId);
+    const createSchedule = this.orgScheduleRepository.create(schedules);
+    return this.orgScheduleRepository.save(createSchedule);
+  }
+
+  async remove(id: number, orgId: number) {
+    await this.checkOwnership([id], orgId);
     return this.orgScheduleRepository.delete(id);
+  }
+
+  async checkOwnership(ids: number[], orgId: number): Promise<boolean> {
+    const schedules = await this.orgScheduleRepository.findBy({ id: In(ids) });
+    return (
+      schedules.length === ids.length &&
+      schedules.every((schedule) => schedule.organization.id === orgId)
+    );
   }
 }
